Example breakdowns!
This is a BIG update for calibration and example assessments.

1. Evaluation plugins now have the option of displaying an explanation
of how they determined a user's score. They do so by exposing a method
called `prepare_explanation_for_assessor`, which returns a renderable
object. The eval plugin must also expose a renderer which can render
this object.

2. A user's example assessments are now available even after the phase
has switched away from the assessment phase. This is in a new page
called `exassessments.php`.