Hide delete button when editing
The delete button looks a lot like an “exit editing” button and is easy
to click instead of the cancel button. The best solution is to hide it
entirely. The whole action bar is now unavailable while editing a
question.

Hitting Cancel before saving a question now deletes it from the
questionnaire.