ID/CMID bug
Apparently Moodle won’t let you insert a record with an ID field. So,
we need to redundantly have an ID primary key AS WELL as a unique CMID
field.

upgrade.php was changed, although the regular upgrade path wouldn’t let
this occur, so the change actually does nothing.