extending question type for reports
Reports need a bit more reflection on the question type, including some
human readable attributes. Whilst a user-supplied title is not
required, you need some way of referring to a question in reports; this
is `title()`. Similarly, since `opinion_of` returns a normalised float
0..1, and since it is undefined when `has_value()` returns false, we
use `opinion_of_readable` for a human-readable version of an opinion in
reports.