Shift to identifying teamevals by ID
Teamevals are now created and fetched by ID, but can be created and
fetched by coursemodule ID or context ID as well. Teamevals can also
have a null cmid; in this case they are templates that live in a
context and are accessed through the teamevaltemplates block. Either
contextid or cmid must be set in a teameval.