Likert now uses fragments for its form
The advantage here is that we can use the rich text editor. The
disadvantage is that editing now requires a server round trip, and
that’s *almost* unavoidable. I may at some point try to build some
standard code for loading the empty form and filling it with the
editing context, but at the moment it round trips.