Fixing messed up TinyMCE bug
For reasons known only to the people who wrote it, the TinyMCE Moodle
plugin doesn’t continuously update its backing store unless you make it
a required field. It’s not a required field for any of the question
plugins, so it doesn’t continuously update its backing store. This new
tiny AMD module fixes that.