Fixed: Likert does not update its client side editing view
Likert questions would fail to update their own editing context client
side, meaning that if a question was created, saved, edited, saved,
then edited again, all without a page refresh, the original data that
was submitted when the question was first created would be visible to
the user, not the updated version. This did not stop edits from being
submitted to the server, but it did stop it from looking right on the
client.