Updated for 3.1.7
So this is an enormous update, and unfortunately for all involved we've had to do quite a lot of development in a private repo. But: this version is much, much faster, and thanks to some changes we've made to our system, it will probably be a lot easier to develop in public once again as we roll into the beta period.

Team Evaluation is still not finished. There is still a lot of work to be done.